cleanup uses of trpc api
Before, there were several methods to get the 'api' variable. It was
(re-)exported in many places, and type inference was tricky. Trpc
client generation has been generalized to a single function that
takes a single type parameter. As a result, all `c.ts`files can now be
removed. Things can be improved further, but this is good for now.