Overhaul plugin system
Previously, it was difficult to load plugins because there was no
central code to load them in. When code needed plugin code, they would
import it (statically or dynamically) and use it directly. This was OK
at the time since the code that used it was exclusively from the main
app. But, now that plugins are using other plugin code, there needs to
be a central place to execute calls. This is the purpose of the
`registerPlugins` and `getPlugins` functions in
`@quasipanacea/common/client/plugin.ts`.

For now, these functions only return the plugin objects, which have to
be handled manually. In the future, this should contain wrapper
functions for each plugin interface.