Rework plugin system
Before, 'view' was a separate concept. Now, both models and pods have
their own view. The non-view component contain the implementation of
handling a particular "mime type". The view component uses this
implementation. This was done to make it easier to make a view layer (so
many lines of backend don't need to be implemented each time).

With each resource and its view, there needs to be a way to connect them.
This is done through "synthetic" mime types like "x-multipart/x-graph".
Later, implement something that more clearly shows what files/data is
handled and how.